+++ Network/HPACK/HeaderBlock/Integer.hs
@@ -38,6 +38,7 @@ if I < 2^N - 1, encode I on N bits
 encodeInteger :: Int -> Int -> IO ByteString
 encodeInteger n i = withTemporaryBuffer 4096 $ \wbuf -> encode wbuf id n i
 
+-- Using writeWord8 is faster than using internals directly.
 {-# INLINABLE encode #-}
 encode :: WorkingBuffer -> (Word8 -> Word8) -> Int -> Int -> IO ()
 encode wbuf set n i